trainer and predictor to old
