+++ clinicadl/commandline/pipelines/train/from_json/cli.py
@@ -6,7 +6,7 @@ from clinicadl.commandline import arguments
 from clinicadl.commandline.modules_options import (
     split,
 )
-from clinicadl.trainer.trainer import Trainer
+from clinicadl.trainer.old_trainer import Trainer
 
 
 @click.command(name="from_json", no_args_is_help=True)
