+++ clinicadl/splitter/old_splitter.py
